Refactored and updated folder structure Added selenium and webdrivermanager dependency created basic UI test
