+++ Pipfile
@@ -6,8 +6,11 @@ name = "pypi"
 [packages]
 requests = "==2.26.0"
 pytest = "==6.2.5"
+assertpy = "*"
+selenium = "*"
+webdriver-manager = "*"
 
 [dev-packages]
 
 [requires]
-python_version = "3.9"
+python_version = "3.10"
