+++ requirements.txt
@@ -1,2 +1,5 @@
 pytest==6.2.5
 requests==2.26.0
+assertpy==1.1
+selenium==4.1.0
+webdriver-manager==3.5.2
