+++ tests/api/booking/createBookingRequest.json
@@ -1,0 +1,11 @@
+{
+  "firstname": "Deepa",
+  "lastname": "Seeba",
+  "totalprice": 111,
+  "depositpaid": true,
+  "bookingdates": {
+    "checkin": "2018-01-01",
+    "checkout": "2019-01-01"
+  },
+  "additionalneeds": "Breakfast"
+}
