+++ tests/api/booking/partialUpdateBookingRequest.json
@@ -1,0 +1,3 @@
+{
+  "firstname": "Superman"
+}
