--- tests/booking/createBookingRequest.json
@@ -1,11 +1,0 @@
-{
-  "firstname": "Deepa",
-  "lastname": "Seeba",
-  "totalprice": 111,
-  "depositpaid": true,
-  "bookingdates": {
-    "checkin": "2018-01-01",
-    "checkout": "2019-01-01"
-  },
-  "additionalneeds": "Breakfast"
-}
