--- tests/booking/partialUpdateBookingRequest.json
@@ -1,3 +1,0 @@
-{
-  "firstname": "Superman"
-}
