--- tests/booking/updateBookingRequest.json
@@ -1,11 +1,0 @@
-{
-  "firstname": "Avishek",
-  "lastname": "Behera",
-  "totalprice": 111,
-  "depositpaid": true,
-  "bookingdates": {
-    "checkin": "2018-01-01",
-    "checkout": "2019-01-01"
-  },
-  "additionalneeds": "Breakfast"
-}
